requirements.txt
Pillow==9.2.0
numpy==1.22.4
#or whatever works for you!